refact: use override for linting ts
We don't want to run all files through the eslint typescript
parser - only the ts files. We are now using the ember-cli
defaults for js files and override the linting settings for
typescript files.

This gets rid of weird warnings in js files were the ts parser
would complain that there are no return types in js files - an
example was the config/deploy.js file.